requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.1.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cvzone==1.6.1
cycler==0.12.1
fastapi==0.115.11
flatbuffers==25.2.10
fonttools==4.56.0
h11==0.14.0
idna==3.10
jax==0.5.2
jaxlib==0.5.1
kiwisolver==1.4.8
matplotlib==3.10.1
mediapipe==0.10.5
ml_dtypes==0.5.1
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==24.2
pillow==11.1.0
protobuf==3.20.3
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.46.1
typing_extensions==4.12.2
uvicorn==0.34.0
